Mundaça do tipo do isbn

+++ src/main/java/br/com/ra/biblioteca_organizator/repositories/BooksRepository.java
@@ -8,7 +8,7 @@ import br.com.ra.biblioteca_organizator.entities.Books;
 
 public interface BooksRepository extends JpaRepository<Books, Integer> {
 
-    Books findByIsbn(Integer isbn);
+    Books findByIsbn(long isbn);
 
     List<Books> findByTitle(String title);
